+++ examples/offscreen_hdr.py
@@ -16,17 +16,17 @@ import png  # provided by the pypng package
 from rendercanvas.offscreen import RenderCanvas
 
 try:
-    from .triangle import setup_drawing_sync
+    from .cube import setup_drawing_sync
 except ImportError:
-    from triangle import setup_drawing_sync
+    from cube import setup_drawing_sync
 
 
 canvas = RenderCanvas(size=(640, 480), pixel_ratio=2)
 draw_frame = setup_drawing_sync(canvas.get_wgpu_context(), format="rgba16float")
 canvas.request_draw(draw_frame)
 
+# Draw, and download image
 image = canvas.draw()
-image = np.asarray(image)
 
 # Convert to RGB
 image = image[:, :, :3]